refactor: extract init.rs from backend/mod.rs
Reduces mod.rs from 1222 to 975 lines.
Part of upstream codebase refactoring (commit 98d215d).
914 tests pass.
